:recycle: (PC-5945) need function in 2 differents pages
